[dev.cc] cmd/asm: rewrite parsing of TEXT $3-2
It was too complicated, assuming the syntax is more general than reality.
It must be a possibly negative integer followed by an optional minus sign
and positive integer. Literals only, no expresssions.

Also put in a TODO about address parsing and clean up a couple of types.

Reviewed-on: https://go-review.googlesource.com/3370

+++ src/cmd/asm/internal/arch/arch.go
@@ -21,8 +21,8 @@ const (
 // Arch wraps the link architecture object with more architecture-specific information.
 type Arch struct {
 	*obj.LinkArch
-	D_INDIR  int16 // TODO: why not in LinkArch?
-	D_CONST2 int16 // TODO: why not in LinkArch?
+	D_INDIR  int // TODO: why not in LinkArch?
+	D_CONST2 int // TODO: why not in LinkArch?
 	// Register number of hardware stack pointer.
 	SP int
 	// Encoding of non-address.
